Allow multiple input files #25 #11
The pattern matching compiler accepted exactly one input file.
In the absence of the `--Input` command line option, an `error`
was thrown. I could have used a `Maybe` value for the `inputFile`
of the `Options` data type and `report` an error when the field
is set to `Nothing`. However, since there are plans to support
multiple input files anyway (see #11), I've decided to use a
list for `inputFiles` instead and show the usage information
when there is no input file (instead of reporting an error).