Simplify `Decl` type #31
Type signatures are not needed by the pattern matching compiler.
Thus, the `TypeSig` constructor of `Decl` could be removed.
Furthermore, the name of the data type is the only information
from the data type declarations's head. Thus, the `DeclHead`
constructor could be removed in favor of a `Name` field in
for the `DataDecl` constructor.