Make `getMatchName` total #38
It is impossible for a `S.FunBind` to have no `S.Match`es.
We could have lifted `getMatchName` into `Sem` and add the
`Report` effect to report an `Internal` error message.
However, `HST.Feature.GuardElimination` uses `head` already
to determine the arity of the function. Thus, I think that
it is safe to use `head` to get the name of the function as
well.