Add selectors for pattern names #38
In `HST.CoreAlgorithm` there are functions for getting the
name of a constructor or variable matched by a pattern.
The functions are used in other files as well and have
incomplete pattern matching. Alternative selectors with
complete pattern matching and proper error reporting have
been added to replace `getQNamePat` and `getPVarName`.