Parameterize HSE front end over source spans
In order to use the source span data type of the Free Compiler,
the HSE front end now allows arbitrary source span annotation
types. The transformer of the HSE front end requires the
source span to implement type classes for the transformation
in our source span representation and back.
The `ParsedModule` and `ParsedExp` type classes had to be
moved from `Parsable` to `Transformable` since not all
source span types have a `Parsable` instance but the
transformation can still be applied to modules with other
source spans as long as the type classes are implemented.
The parser and pretty printer now also perform the transformation
to and from our syntax internally.